refactor(cli): import the split parser modules directly
Follows the docs-index split on the base branch: the renderer reads the
link scanner from docs-links and the payload builder reads the
frontmatter parser from docs-frontmatter, instead of going through the
facade.

Refs: RFC 0005
