+++ packages/cli/src/docs-viewer.ts
@@ -10,7 +10,9 @@
 import { readFile } from 'node:fs/promises'
 import { fileURLToPath } from 'node:url'
 import { resolve } from 'node:path'
-import { localLinkTarget, parseDocFrontmatter, scanDocs, type DocActorEvent, type DocRef } from './docs-index'
+import { parseDocFrontmatter } from './docs-frontmatter'
+import { localLinkTarget } from './docs-links'
+import { scanDocs, type DocActorEvent, type DocRef } from './docs-index'
 import { runDocsCheck, resolveDocLink } from './docs-check'
 import { buildDocsGraph, type DocsGraphEdge, type DocsGraphNode } from './docs-graph'
 import { renderDocHtml } from './docs-render'
